Rename device_impl.{h|cpp} to device.{h|cpp}
