+++ external_example/plugins/example/example_impl.cpp
@@ -1,5 +1,5 @@
 #include "example_impl.h"
-#include "device_impl.h"
+#include "device.h"
 #include "global_include.h"
 #include <functional>
